Re-organize project to avoid duplication
