+++ golang/.dockerignore
@@ -1,0 +1,4 @@
+*.yml
+*.json
+.gitignore
+.env
