--- golang/backend/go.mod
@@ -1,10 +1,0 @@
-module github.com/wihrt/idle_arena/arena
-
-go 1.16
-
-require (
-	github.com/gorilla/mux v1.8.0
-	github.com/levigross/grequests v0.0.0-20190908174114-253788527a1a
-	go.mongodb.org/mongo-driver v1.7.0
-	go.uber.org/zap v1.18.1
-)
