+++ golang/dice/dice.go
