+++ golang/gladiator/name.go
